#1434 Rename reducres to reducers
