+++ IsraelHiking.Web/src/application/components/dialogs/private-poi-show-dialog.component.ts
@@ -8,7 +8,7 @@ import { ImageGalleryService } from "../../services/image-gallery.service";
 import { PrivatePoiEditDialogComponent } from "./private-poi-edit-dialog.component";
 import { AddSimplePoiDialogComponent } from "./add-simple-poi-dialog.component";
 import { SelectedRouteService } from "../../services/layers/routelayers/selected-route.service";
-import { AddPrivatePoiAction } from "../../reducres/routes.reducer";
+import { AddPrivatePoiAction } from "../../reducers/routes.reducer";
 import { ApplicationState, MarkerData, LinkData } from "../../models/models";
 
 interface IPrivatePoiShowDialogData {
