+++ IsraelHiking.Web/src/application/components/intercations/route-edit-poi.interaction.ts
@@ -3,7 +3,7 @@ import { MapMouseEvent, Map } from "mapbox-gl";
 import { MatDialog } from "@angular/material/dialog";
 import { NgRedux } from "@angular-redux/store";
 
-import { AddPrivatePoiAction, UpdatePrivatePoiAction } from "../../reducres/routes.reducer";
+import { AddPrivatePoiAction, UpdatePrivatePoiAction } from "../../reducers/routes.reducer";
 import { SelectedRouteService } from "../../services/layers/routelayers/selected-route.service";
 import { PrivatePoiEditDialogComponent } from "../dialogs/private-poi-edit-dialog.component";
 import { GeoLocationService } from "../../services/geo-location.service";
