+++ IsraelHiking.Web/src/application/components/main-menu.component.ts
@@ -22,8 +22,8 @@ import { SharesDialogComponent } from "./dialogs/shares-dialog.component";
 import { ConfigurationDialogComponent } from "./dialogs/configuration-dialog.component";
 import { LanguageDialogComponent } from "./dialogs/language-dialog.component";
 import { FilesSharesDialogComponent } from "./dialogs/files-shares-dialog.component";
-import { SetUIComponentVisibilityAction } from "../reducres/ui-components.reducer";
-import { SetAgreeToTermsAction } from "../reducres/user.reducer";
+import { SetUIComponentVisibilityAction } from "../reducers/ui-components.reducer";
+import { SetAgreeToTermsAction } from "../reducers/user.reducer";
 import { UserInfo, ApplicationState } from "../models/models";
 
 @Component({
