+++ IsraelHiking.Web/src/application/components/map/traces.component.ts
@@ -5,8 +5,8 @@ import { Observable } from "rxjs";
 import { BaseMapComponent } from "../base-map.component";
 import { ResourcesService } from "../../services/resources.service";
 import { SpatialService } from "../../services/spatial.service";
-import { RemoveMissingPartAction, SetVisibleTraceAction, SetMissingPartsAction } from "../../reducres/traces.reducer";
-import { AddRouteAction } from "../../reducres/routes.reducer";
+import { RemoveMissingPartAction, SetVisibleTraceAction, SetMissingPartsAction } from "../../reducers/traces.reducer";
+import { AddRouteAction } from "../../reducers/routes.reducer";
 import { RoutesFactory } from "../../services/layers/routelayers/routes.factory";
 import { Trace, ApplicationState, LatLngAlt } from "../../models/models";
