+++ IsraelHiking.Web/src/application/components/overlays/gps-location-overlay.component.ts
@@ -11,8 +11,8 @@ import { ResourcesService } from "../../services/resources.service";
 import { SelectedRouteService } from "../../services/layers/routelayers/selected-route.service";
 import { RunningContextService } from "../../services/running-context.service";
 import { HashService } from "../../services/hash.service";
-import { AddPrivatePoiAction } from "../../reducres/routes.reducer";
-import { ToggleDistanceAction } from "../../reducres/in-memory.reducer";
+import { AddPrivatePoiAction } from "../../reducers/routes.reducer";
+import { ToggleDistanceAction } from "../../reducers/in-memory.reducer";
 import { ApplicationState, LatLngAlt } from "../../models/models";
 
 @Component({
