+++ IsraelHiking.Web/src/application/components/sidebar/publicpoi/public-poi-sidebar.component.ts
@@ -11,10 +11,10 @@ import { AuthorizationService } from "../../../services/authorization.service";
 import { ToastService } from "../../../services/toast.service";
 import { HashService, RouteStrings, IPoiRouterData } from "../../../services/hash.service";
 import { SelectedRouteService } from "../../../services/layers/routelayers/selected-route.service";
-import { AddRouteAction, AddPrivatePoiAction } from "../../../reducres/routes.reducer";
+import { AddRouteAction, AddPrivatePoiAction } from "../../../reducers/routes.reducer";
 import { RoutesFactory } from "../../../services/layers/routelayers/routes.factory";
 import { FitBoundsService } from "../../../services/fit-bounds.service";
-import { SetSelectedPoiAction, SetUploadMarkerDataAction, SetSidebarAction } from "../../../reducres/poi.reducer";
+import { SetSelectedPoiAction, SetUploadMarkerDataAction, SetSidebarAction } from "../../../reducers/poi.reducer";
 import { SpatialService } from "../../../services/spatial.service";
 import { RunningContextService } from "../../../services/running-context.service";
 import { SidebarService } from "../../../services/sidebar.service";
