+++ IsraelHiking.Web/src/application/reducers/configuration.reducer.ts
