+++ IsraelHiking.Web/src/application/reducers/in-memory.reducer.ts
