+++ IsraelHiking.Web/src/application/reducers/initial-state.ts
