+++ IsraelHiking.Web/src/application/reducers/layers.reducer.ts
