+++ IsraelHiking.Web/src/application/reducers/location.reducer.ts
