+++ IsraelHiking.Web/src/application/reducers/offline.reducer.ts
