+++ IsraelHiking.Web/src/application/reducers/poi.reducer.ts
