+++ IsraelHiking.Web/src/application/reducers/root.reducer.ts
