+++ IsraelHiking.Web/src/application/reducers/route-editing-state.reducer.ts
