+++ IsraelHiking.Web/src/application/reducers/routes.reducer.ts
