+++ IsraelHiking.Web/src/application/reducers/share-urls.reducer.ts
