+++ IsraelHiking.Web/src/application/reducers/traces.reducer.ts
