+++ IsraelHiking.Web/src/application/reducers/ui-components.reducer.ts
