+++ IsraelHiking.Web/src/application/reducers/user.reducer.ts
