+++ IsraelHiking.Web/src/application/services/geo-location.service.ts
@@ -13,7 +13,7 @@ import { ResourcesService } from "./resources.service";
 import { RunningContextService } from "./running-context.service";
 import { LoggingService } from "./logging.service";
 import { ToastService } from "./toast.service";
-import { SetGeoLocationStateAction } from "../reducres/in-memory.reducer";
+import { SetGeoLocationStateAction } from "../reducers/in-memory.reducer";
 import { ApplicationState, ILatLngTime } from "../models/models";
 
 @Injectable()
