+++ IsraelHiking.Web/src/application/services/layers/routelayers/selected-route.service.ts
@@ -3,7 +3,7 @@ import { NgRedux, select } from "@angular-redux/store";
 import { Observable } from "rxjs";
 import { some } from "lodash-es";
 
-import { SetSelectedRouteAction } from "../../../reducres/route-editing-state.reducer";
+import { SetSelectedRouteAction } from "../../../reducers/route-editing-state.reducer";
 import {
     AddRouteAction,
     SplitRouteAction,
@@ -14,7 +14,7 @@ import {
     ReplaceSegmentsAction,
     AddPrivatePoiAction,
     ChangeEditStateAction
-} from "../../../reducres/routes.reducer";
+} from "../../../reducers/routes.reducer";
 import { RoutesFactory } from "./routes.factory";
 import { ResourcesService } from "../../resources.service";
 import { SpatialService } from "../../spatial.service";
