+++ IsraelHiking.Web/src/application/services/resources.service.ts
@@ -3,7 +3,7 @@ import { Direction } from "@angular/cdk/bidi";
 import { NgRedux } from "@angular-redux/store";
 
 import { GetTextCatalogService } from "./gettext-catalog.service";
-import { SetLanguageAction } from "../reducres/configuration.reducer";
+import { SetLanguageAction } from "../reducers/configuration.reducer";
 import { ApplicationState, Language } from "../models/models";
 import { Urls } from "../urls";
